Extract common validation logic to a trait
